requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
hatch-rest-api>=1.20.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
apscheduler>=3.10.0
redis>=5.0.0
